Renaming template on views
