+++ src/views/create-trip-events-list-wrapper.js
