+++ src/views/create-trip-events-list.js
