+++ src/views/event-item/create-trip-event-item-edit/create-event-item-collapsible-btn.js
