+++ src/views/event-item/create-trip-event-item-edit/create-trip-event-details-wrapper.js
