+++ src/views/event-item/create-trip-event-item-edit/create-trip-event-item-edit-offers.js
