+++ src/views/event-item/create-trip-event-item.js
