+++ src/views/header/create-trip-info-header.js
