+++ src/views/header/create-trip-tabs-header.js
